requirements.txt
# pip install -r requirements.txt
Flask
opencv-python
Werkzeug
paramiko
requests
pytz
python-telegram-bot
python-dotenv